requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-request==1.6.1
idna==3.3
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9

